debezium/dbz#1808 Remove the useConnectorForm and move the type content
